project/requirements.txt
Flask==2.1.1
torch==1.12.0
Flask-WTF==1.0.1
opencv-python==4.6.0.66
Pillow==7.2.0
